+++ engine/engine/rope_manager.h
@@ -20,7 +20,7 @@ public:
 	void clearStrings();
 	static CRope_manager& get();
 
-	int getRopeCount() { return strings.size(); }
+	int getRopeCount() { return (int)strings.size(); }
 
 	std::deque<CHandle> getStrings() { return strings; }
 };
